chore(hosting): move hosting dir
