+++ hosting/.editorconfig
