+++ hosting/package.json
