+++ hosting/src/components/pages/Top/TopContainer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { withRouter, RouteComponentProps } from 'react-router';
 import { ThemeContainer } from '../../../stores/theme';
 import { UserContainer } from '../../../stores/user';
-import { Top, TopProps } from './';
+import { Top, TopProps } from '.';
 
 interface Props extends RouteComponentProps, Partial<TopProps> {}
