+++ hosting/src/constants/GlobalStyle.tsx
