+++ hosting/src/serviceWorker.ts
